moved all simulation code to simulator directory
