+++ case_studies/summer_template/config.yaml
@@ -17,9 +17,8 @@ paths:
     output: ${paths.project}/output
 
 
-# TODO: add the ability to use simulate datasets that have been stored on disk
-# TODO: make whether datasets include labels or not that can be used for training
-# TODO: make DC2 dataset with labels and training on it
+# TODO: add the ability to use simulated datasets that have been stored on disk
+# TODO: make DC2 dataset/simulator with labels and train with it
 # TODO: create some multiband datasets
 # TODO: create configs files like this one for DECaLS and DC2
 
@@ -36,7 +35,7 @@ simulator:
         alpha: 0.43
         prob_galaxy: 0.30
         galaxy_prior:
-            _target_: bliss.datasets.galsim_galaxies.SingleGalsimGalaxyPrior
+            _target_: bliss.simulator.galsim_galaxies.SingleGalsimGalaxyPrior
             flux_sample: "pareto"
             min_flux: 622.0
             max_flux: 1e6
@@ -57,7 +56,7 @@ simulator:
             - 2
         psf_params_file: ${paths.data}/sdss/94/1/12/psField-000094-1-0012.fits
         galaxy_model:
-            _target_: bliss.datasets.galsim_galaxies.SingleGalsimGalaxyDecoder
+            _target_: bliss.simulator.galsim_galaxies.SingleGalsimGalaxyDecoder
             n_bands: 1
             slen: 53
             pixel_scale: 0.396
@@ -66,7 +65,7 @@ simulator:
             sdss_bands:
                 - 2
     background:
-        _target_: bliss.datasets.background.SimulatedSDSSBackground
+        _target_: bliss.simulator.background.SimulatedSDSSBackground
         sdss_dir: ${paths.sdss}
         run: 94
         camcol: 1
@@ -126,11 +125,12 @@ encoder:
             [-1, 3, C3, [256, false]],
             [[17], 1, Detect, [nc, anchors]],
         ]
+    # TODO: implement pretraining and try pretraining with crowded fields
+    # pretrained_weights: ${paths.project}/models/${training.name}.pt
+
 
 training:
     name: "sdss_encoder"
-    dataset: ${simulator}
-    encoder: ${encoder}
     version: null
     n_epochs: 1000
     experiment: default
